news-digest-llm/requirements.txt
openai==1.56.2
feedparser
python-dotenv
